Se muestran las secciones segun los permisos del usuario
