+++ src/pages/home/home.ts
@@ -6,6 +6,7 @@ import { LoginPage } from '../login/login';
 import { NativeAudio } from '@ionic-native/native-audio';
 import { PushService } from '../../services/push.service';
 import { Platform } from 'ionic-angular';
+import { AngularFireDatabase } from 'angularfire2/database';
 
 @Component({
   selector: 'page-home',
@@ -21,10 +22,20 @@ export class HomePage {
     public authAf: AngularFireAuth,
     private alertCtrl: AlertController,
     private nativeAudio: NativeAudio,
+    public af: AngularFireDatabase,
     private platform: Platform,
-    private pushService: PushService) {
-    var pagesService = new PagesService();
-    this.pages = pagesService.getByUserType(this.user);
+    private pushService: PushService,
+    private pagesService: PagesService) {
+    af.database.ref("/usuarios/").on('value', usuarios => {
+        let props = Object.getOwnPropertyNames(usuarios.val());
+        props.forEach(p => {
+            let usr = usuarios.val()[p];
+            if(usr.email == this.authAf.auth.currentUser.email){
+              alert(usr.tipo);
+              this.pages = pagesService.getByUserType(usr.tipo);
+            }
+        });
+    });
     this.pushService.getPermission(this.platform);
     this.pushService.receiveMessage(this.platform);
   }
